Updated parsing handler for larger data

+++ src/Models/CreateFile.php
@@ -16,6 +16,27 @@ class CreateFile
      */
     public function unformatted($filename, $contents)
     {
+        if(file_exists($filename)){
+
+            return false;
+
+        }
         return file_put_contents($filename, $contents);
+    }
+
+    public function json($filename, $contents)
+    {
+        $contents = json_encode($contents);
+
+        return self::unformatted($filename, $contents);
+    }
+
+    public function directory($filepath)
+    {
+        if(is_dir($filepath)){
+            return false;
+        }
+
+        return mkdir($filepath, 0777, true);
     }
 }
